bot/requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.3
pytgcalls==1.0.5
ntgcalls==1.3.1
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
aiohttp==3.11.11
aiofiles==24.1.0
yt-dlp==2024.12.23
